move mock folder to test and update files
