--- idea/data-storage/src/common/mock/code/collection-code.mock.yaml
@@ -1,21 +1,0 @@
-collectionCode:
-  - id: '0x7350'
-    genesis: '0x7350'
-    status: 'Active'
-    name: 'code_1'
-    blockHash: 'code_1'
-    timestamp: NULL
-
-  - id: '0x7351'
-    genesis: '0x7351'
-    status: 'Inactive'
-    name: 'code_2'
-    blockHash: 'code_2'
-    timestamp: NULL
-
-  - id: '0x7352'
-    genesis: '0x7352'
-    status: 'Active'
-    name: 'code_3'
-    blockHash: 'code_3'
-    timestamp: NULL
