--- idea/data-storage/src/common/mock/message/messages.mock.yaml
@@ -1,21 +1,0 @@
-Messages:
-  - id: '0x7350'
-    genesis: '0x7350'
-    processedWithPanic: false
-    destination: 'message_1'
-    payload: 'message_1'
-    readStatus: NULL
-
-  - id: '0x7351'
-    genesis: '0x7351'
-    processedWithPanic: true
-    destination: 'message_2'
-    payload: 'message_2'
-    readStatus: NULL
-
-  - id: '0x7352'
-    genesis: '0x7352'
-    processedWithPanic: true
-    destination: 'message_3'
-    payload: 'message_3'
-    readStatus: NULL
