--- idea/data-storage/src/common/mock/metadata/collection-metadata.mock.yaml
@@ -1,18 +1,0 @@
-collectionMetadata:
-  - id: '0x7350'
-    owner: 'metadata_1'
-    metaFile: 'metadata_1'
-    meta: 'metadata_1'
-    program: 'metadata_1'
-
-  - id: '0x7351'
-    owner: 'metadata_2'
-    metaFile: 'metadata_2'
-    meta: 'metadata_2'
-    program: 'metadata_2'
-
-  - id: '0x7352'
-    owner: 'metadata_3'
-    metaFile: 'metadata_3'
-    meta: 'metadata_3'
-    program: 'metadata_3'
